TG-27: Fix comments
