+++ src/Components/Tasks/row/index.js
@@ -1,25 +1,26 @@
 import React from 'react';
 
-function Row(props) {
+const Row = (props) => {
   const { task, deleteTask } = props;
   return (
     <tr id={task._id} className={StyleSheet.row}>
       <td>{task._id}</td>
       <td>{task.description}</td>
       <td>
-        <i
-          className="fa-solid fa-xmark"
+        <a
           onClick={() => {
             deleteTask(task._id);
           }}
-        ></i>
+        >
+          <button>X</button>
+        </a>
       </td>
       <td>
         <a href={`http://localhost:3000/task/${task._id}`}>
-          <button>Edit</button>{' '}
+          <button>Edit</button>
         </a>
       </td>
     </tr>
   );
-}
+};
 export default Row;
